Move files to profile/actions
